Fixup relationships and entity states tables/models

+++ database/migrations/2019_01_01_231440_create_comments_table.php
@@ -17,7 +17,7 @@ class CreateCommentsTable extends Migration
             $table->bigIncrements('id');
             $table->uuid('uuid')->unique();
             $table->text('body');
-            $table->unsignedInteger('commentable_id');
+            $table->unsignedBigInteger('commentable_id');
             $table->string('commentable_type');
             $table->timestamps();
         });
